C-7

+++ C-7/ex02/Array.hpp
@@ -9,7 +9,7 @@ template <typename T>
 class Array{
     private:
         T *_array;
-        unsigned int _len;
+        int _len;
     public:
 
         class InvalidSize: public std::exception{
@@ -21,13 +21,13 @@ class Array{
 
         Array(): _array(NULL), _len(0){}
         
-        Array(unsigned int len): _len(len){
+        Array(int len): _len(len){
             this->_array = new T[this->_len];
         }
 
         Array(const Array &copy): _len(copy._len){
             this->_array = new T[this->_len];
-            for (unsigned int i = 0; i < _len; i++)
+            for (int i = 0; i < _len; i++)
                 this->_array[i] = copy._array[i];
         }
 
@@ -41,7 +41,7 @@ class Array{
                     this->_array = NULL;
                 else{
                     this->array = new T[this->_len]();
-                    for (unsigned int i = 0; i < _len; i++){
+                    for (int i = 0; i < _len; i++){
                         _array[i] = copy._array[i];
                     }
                 }
@@ -50,13 +50,13 @@ class Array{
         }
 
         T   &operator[](int index){
-            if (index < 0 || (unsigned)index >= _len)
+            if (index < 0 || index >= _len)
                 throw InvalidSize();
             return _array[index];
         }
 
         T   &operator[](int index)const{
-            if (index < 0 || (unsigned)index >= _len)
+            if (index < 0 || index >= _len)
                 throw InvalidSize();
             return _array[index];
         }
